cleanup stream/backend organization
